Refactor🔨: divide component
struct 분리.
