+++ EcoTales/EcoTales/Screens/Chapter3/Views/FindWrongGameView.swift
@@ -94,109 +94,16 @@ private struct UpperPanelView: View {
 }
 
 private struct ExplainGameView: View {
+
     private var owlLine = ["콜록! 콜록! 공장 매연은 나빠!", "공장 매연은 동물들에게 나빠!", "직구 배송을 주의하자!", "대중교통을 이용하자!", "자동차는 되도록 자제하자!"]
     @EnvironmentObject var chapterThreeUserValue: ChapterThreeUserSettings
+
     var body: some View {
         ZStack {
             Text(owlLine[chapterThreeUserValue.circleScore])
                 .background(Image("chapter1_turtleDialog"))
         }
         .frame(width: 450, height: 30)
-    }
-}
-
-private struct HeartTake: View {
-    @EnvironmentObject var chapterThreeUserValue: ChapterThreeUserSettings
-    var body: some View {
-        ZStack {
-            HStack {
-                Spacer()
-                Image("findWrong_heartEmpty")
-                    .resizable()
-                    .frame(width: 30, height: 30)
-                Image("findWrong_heartEmpty")
-                    .resizable()
-                    .frame(width: 30, height: 30)
-                Image("findWrong_heartEmpty")
-                    .resizable()
-                    .frame(width: 30, height: 30)
-            }
-            HStack {
-                heartTake( chapterThreeUserValue.lifeCount )
-            }
-        }
-    }
-    @ViewBuilder func heartTake( _ lifeCount: Int) -> some View {
-        HStack {
-            Spacer()
-            if lifeCount > 0 {
-                ForEach(1...lifeCount, id: \.self) { _ in
-                    Image("findWrong_heartFill")
-                        .resizable()
-                        .frame(width: 30, height: 30)
-                }
-            }
-        }
-    }
-}
-
-private struct FindWrongScreenView: View {
-    @EnvironmentObject var chapterThreeUserValue: ChapterThreeUserSettings
-    @Binding var isFindWrongClear: Bool
-    var body: some View {
-        ZStack {
-            Image("chapter3_GameBackground")
-                .resizable()
-                .onTapGesture {
-                    chapterThreeUserValue.lifeCount -= 1
-                }
-            findWrongButton(0, firstCircle: CGPoint(x: 405, y: 198),
-                            secondCircle: CGPoint(x: 30, y: 198), circleSize: 140)
-            findWrongButton(1, firstCircle: CGPoint(x: 580, y: 140),
-                            secondCircle: CGPoint(x: 210, y: 140), circleSize: 120)
-            findWrongButton(2, firstCircle: CGPoint(x: 505, y: 20),
-                            secondCircle: CGPoint(x: 135, y: 20), circleSize: 60)
-            findWrongButton(3, firstCircle: CGPoint(x: 520, y: 280),
-                            secondCircle: CGPoint(x: 150, y: 280), circleSize: 70)
-            findWrongButton(4, firstCircle: CGPoint(x: 570, y: 260),
-                            secondCircle: CGPoint(x: 200, y: 260), circleSize: 70)
-        }
-    }
-    @ViewBuilder func findWrongButton( _ circleNum: Int, firstCircle: CGPoint,
-                                       secondCircle: CGPoint, circleSize: CGFloat
-    ) -> some View {
-        Button {
-            chapterThreeUserValue.isShowCircle[circleNum] = true
-            chapterThreeUserValue.circleScore = circleNum
-            let sum = chapterThreeUserValue.isShowCircle.reduce(true) {
-                $0 && $1
-            }
-            if sum {
-                isFindWrongClear = true
-            }
-        } label: {
-            Image(systemName: "circle")
-                .resizable()
-                .frame(width: circleSize, height: circleSize)
-                .opacity(chapterThreeUserValue.isShowCircle[circleNum] ? 1.0 : 0.0)
-        }
-        .position(x: firstCircle.x, y: firstCircle.y)
-        Button {
-            chapterThreeUserValue.isShowCircle[circleNum] = true
-            chapterThreeUserValue.circleScore = circleNum
-            let sum = chapterThreeUserValue.isShowCircle.reduce(true) {
-                $0 && $1
-            }
-            if sum {
-                isFindWrongClear = true
-            }
-        } label: {
-            Image(systemName: "circle")
-                .resizable()
-                .frame(width: circleSize, height: circleSize)
-                .opacity(chapterThreeUserValue.isShowCircle[circleNum] ? 1.0 : 0.0)
-        }
-        .position(x: secondCircle.x, y: secondCircle.y)
     }
 }
